+++ NEWS.md
@@ -1,5 +1,9 @@
 # DNH4 0.1.8
 
+* 다음 뉴스의 댓글 구조가 바뀌어서 `getComment` 함수를 대응 개발 하였습니다. 
+
+* Fix `getComment` function to change site.
+
 # DNH4 0.1.7
 
 # DNH4 0.1.6